requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
pgvector==0.3.5
sentence-transformers==5.1.0
ffmpeg-python==0.2.0
aiofiles==24.1.0
requests==2.32.5
pypdf==5.4.0
python-pptx==1.0.2
